requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
decorator==5.2.1
distro==1.9.0
executing==2.2.1
filetype==1.2.0
gitdb==4.0.12
GitPython==3.1.43
google-ai-generativelanguage==0.6.18
google-api-core==2.30.3
google-auth==2.49.2
googleapis-common-protos==1.74.0
groq==0.37.1
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==9.10.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-core==0.3.84
langchain-google-genai==2.1.4
langchain-groq==0.2.1
langsmith==0.7.30
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
narwhals==2.19.0
networkx==3.6.1
numpy==2.4.4
orjson==3.11.8
packaging==24.2
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==5.29.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
pydeck==0.9.1
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
pyvis==0.3.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.40.1
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
wcwidth==0.6.0
xxhash==3.6.0
yt-dlp==2026.3.17
zstandard==0.25.0
